requirements.txt
pygame>=2.5.0
# RPi.GPIO>=0.7.1  # Raspberry Pi specific package
Flask==2.2.3
Werkzeug==2.2.3
python-dotenv==0.19.0
mcrcon==0.7.0
flask-cors==3.0.10
opencv-python>=4.7.0
numpy>=1.23.0
requests>=2.28.0 